// src/state-machine/conversation-state.enum.ts
export enum ConversationState {
  NEW = 'NEW',
  TECHFIND_GREETING = 'TECHFIND_GREETING',
  INDUSTRY_DISCOVERY = 'INDUSTRY_DISCOVERY',
  DEMO_SELECTED = 'DEMO_SELECTED',
  DEMO_RUNNING = 'DEMO_RUNNING',
  DEMO_TRANSACTION = 'DEMO_TRANSACTION',
  VALUE_REVEAL = 'VALUE_REVEAL',
  BUSINESS_QUALIFICATION = 'BUSINESS_QUALIFICATION',
  LEAD_SCORED = 'LEAD_SCORED',
  MEETING_OFFERED = 'MEETING_OFFERED',
  MEETING_BOOKED = 'MEETING_BOOKED',
  HUMAN_HANDOFF = 'HUMAN_HANDOFF',
}

export const CONVERSATION_STATES = Object.values(ConversationState);